Cleaning up template.
