+++ tmpl8/template/precomp.h
@@ -4,7 +4,7 @@
 
 // A precompiled header speeds up compilation by precompiling stable code.
 // The content in this file is source code that you will likely not change
-// for your project. More info: 
+// for your project. More info:
 // www.codeproject.com/Articles/1188975/How-to-Optimize-Compilation-Times-with-Precompil
 
 // common C++ headers
@@ -29,7 +29,7 @@ typedef unsigned int uint;
 typedef unsigned short ushort;
 typedef unsigned short half;
 
-// "leak" common namespaces to all compilation units. This is not standard // C++ practice 
+// "leak" common namespaces to all compilation units. This is not standard // C++ practice
 // but a deliberate simplification for template projects. Feel free to remove this if it
 // offends you.
 using namespace std;
@@ -117,12 +117,6 @@ using namespace Tmpl8;
 #include <unistd.h>
 #endif
 
-// OpenCL headers
-// #define CL_USE_DEPRECATED_OPENCL_2_0_APIS // safe; see https://stackoverflow.com/a/28500846
-#define CL_TARGET_OPENCL_VERSION 300
-#include "cl/cl.h"
-#include <cl/cl_gl.h>
-
 // GLFW
 #define GLFW_USE_CHDIR 0
 #define GLFW_EXPOSE_NATIVE_WIN32
@@ -134,8 +128,7 @@ using namespace Tmpl8;
 // zlib
 #include "zlib.h"
 
-// opencl & opencl
-#include "opencl.h"
+// opengl
 #include "opengl.h"
 
 // fatal error reporting (with a pretty window)
